Fixed permissions bug due to instantiating the start screen twice
