+++ lib/presentation/create_new_alarm_screen/time_pickers/time_wheel_picker.dart
@@ -60,7 +60,7 @@ class _TimeWheelPickerState extends State<TimeWheelPicker> {
               child: Container(
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(8.0),
-                  color: Color.fromARGB(255, 211, 222, 227),
+                  color: const Color.fromARGB(255, 211, 222, 227),
                 ),
                 width: MediaQuery.of(context).size.width,
                 height: 50,
